+++ relay-server/src/services/processor.rs
@@ -74,7 +74,7 @@ use crate::service::ServiceError;
 use crate::services::global_config::GlobalConfigHandle;
 use crate::services::outcome::{DiscardReason, Outcome, TrackOutcome};
 use crate::services::processor::event::FiltersStatus;
-use crate::services::project::ProjectInfo;
+use crate::services::project::ProjectState;
 use crate::services::project_cache::{
     AddMetricBuckets, AddMetricMeta, BucketSource, ProjectCache, UpdateRateLimits,
 };
@@ -84,7 +84,7 @@ use crate::services::upstream::{
 };
 use crate::statsd::{RelayCounters, RelayHistograms, RelayTimers};
 use crate::utils::{
-    self, InvalidProcessingGroupType, ManagedEnvelope, SamplingResult, TypedEnvelope,
+    self, InvalidProcessingGroupType, ManagedEnvelope, SamplingResult, ThreadPool, TypedEnvelope,
 };
 use crate::{http, metrics};
 
@@ -169,7 +169,7 @@ pub trait EventProcessing {}
 /// A trait for processing groups that can be dynamically sampled.
 pub trait Sampling {
     /// Whether dynamic sampling should run under the given project's conditions.
-    fn supports_sampling(project_state: &ProjectInfo) -> bool;
+    fn supports_sampling(project_state: &ProjectState) -> bool;
 
     /// Whether reservoir sampling applies to this processing group (a.k.a. data type).
     fn supports_reservoir_sampling() -> bool;
@@ -179,7 +179,7 @@ processing_group!(TransactionGroup, Transaction);
 impl EventProcessing for TransactionGroup {}
 
 impl Sampling for TransactionGroup {
-    fn supports_sampling(project_state: &ProjectInfo) -> bool {
+    fn supports_sampling(project_state: &ProjectState) -> bool {
         // For transactions, we require transaction metrics to be enabled before sampling.
         matches!(&project_state.config.transaction_metrics, Some(ErrorBoundary::Ok(c)) if c.is_enabled())
     }
@@ -200,7 +200,7 @@ processing_group!(CheckInGroup, CheckIn);
 processing_group!(SpanGroup, Span);
 
 impl Sampling for SpanGroup {
-    fn supports_sampling(project_state: &ProjectInfo) -> bool {
+    fn supports_sampling(project_state: &ProjectState) -> bool {
         // If no metrics could be extracted, do not sample anything.
         matches!(&project_state.config().metric_extraction, ErrorBoundary::Ok(c) if c.is_supported())
     }
@@ -589,14 +589,14 @@ type ExtractedEvent = (Annotated<Event>, usize);
 #[derive(Debug)]
 pub struct ProcessingExtractedMetrics {
     metrics: ExtractedMetrics,
-    project: Arc<ProjectInfo>,
+    project: Arc<ProjectState>,
     global: Arc<GlobalConfig>,
     extrapolation_factor: Option<FiniteF64>,
 }
 
 impl ProcessingExtractedMetrics {
     pub fn new(
-        project: Arc<ProjectInfo>,
+        project: Arc<ProjectState>,
         global: Arc<GlobalConfig>,
         dsc: Option<&DynamicSamplingContext>,
     ) -> Self {
@@ -747,7 +747,7 @@ fn send_metrics(metrics: ExtractedMetrics, envelope: &Envelope, project_cache: A
         // project_without_tracing         -> metrics goes to self
         // dependent_project_with_tracing  -> metrics goes to root
         // root_project_with_tracing       -> metrics goes to root == self
-        let sampling_project_key = envelope.sampling_key().unwrap_or(project_key);
+        let sampling_project_key = utils::get_sampling_key(envelope).unwrap_or(project_key);
         project_cache.send(AddMetricBuckets::internal(
             sampling_project_key,
             sampling_metrics,
@@ -792,11 +792,11 @@ struct ProcessEnvelopeState<'a, Group> {
     extracted_metrics: ProcessingExtractedMetrics,
 
     /// The state of the project that this envelope belongs to.
-    project_state: Arc<ProjectInfo>,
+    project_state: Arc<ProjectState>,
 
     /// The state of the project that initiated the current trace.
     /// This is the config used for trace-based dynamic sampling.
-    sampling_project_state: Option<Arc<ProjectInfo>>,
+    sampling_project_state: Option<Arc<ProjectState>>,
 
     /// The id of the project that this envelope is ingested into.
     ///
@@ -891,8 +891,8 @@ pub struct ProcessEnvelopeResponse {
 #[derive(Debug)]
 pub struct ProcessEnvelope {
     pub envelope: ManagedEnvelope,
-    pub project_info: Arc<ProjectInfo>,
-    pub sampling_project_info: Option<Arc<ProjectInfo>>,
+    pub project_state: Arc<ProjectState>,
+    pub sampling_project_state: Option<Arc<ProjectState>>,
     pub reservoir_counters: ReservoirCounters,
 }
 
@@ -948,8 +948,8 @@ pub struct ProcessMetricMeta {
 pub struct ProjectMetrics {
     /// The metric buckets to encode.
     pub buckets: Vec<Bucket>,
-    /// Project info for extracting quotas.
-    pub project_info: Arc<ProjectInfo>,
+    /// Project state for extracting quotas.
+    pub project_state: Arc<ProjectState>,
 }
 
 /// Encodes metrics into an envelope ready to be sent upstream.
@@ -1113,6 +1113,7 @@ impl Default for Addrs {
 }
 
 struct InnerProcessor {
+    pool: ThreadPool,
     config: Arc<Config>,
     global_config: GlobalConfigHandle,
     cogs: Cogs,
@@ -1132,6 +1133,7 @@ struct InnerProcessor {
 impl EnvelopeProcessorService {
     /// Creates a multi-threaded envelope processor.
     pub fn new(
+        pool: ThreadPool,
         config: Arc<Config>,
         global_config: GlobalConfigHandle,
         cogs: Cogs,
@@ -1150,6 +1152,7 @@ impl EnvelopeProcessorService {
         });
 
         let inner = InnerProcessor {
+            pool,
             global_config,
             cogs,
             #[cfg(feature = "processing")]
@@ -1220,8 +1223,8 @@ impl EnvelopeProcessorService {
         global_config: Arc<GlobalConfig>,
         mut managed_envelope: TypedEnvelope<G>,
         project_id: ProjectId,
-        project_state: Arc<ProjectInfo>,
-        sampling_project_state: Option<Arc<ProjectInfo>>,
+        project_state: Arc<ProjectState>,
+        sampling_project_state: Option<Arc<ProjectState>>,
         reservoir_counters: Arc<Mutex<BTreeMap<RuleId, i64>>>,
     ) -> ProcessEnvelopeState<G> {
         let envelope = managed_envelope.envelope_mut();
@@ -1840,8 +1843,8 @@ impl EnvelopeProcessorService {
         &self,
         mut managed_envelope: ManagedEnvelope,
         project_id: ProjectId,
-        project_state: Arc<ProjectInfo>,
-        sampling_project_state: Option<Arc<ProjectInfo>>,
+        project_state: Arc<ProjectState>,
+        sampling_project_state: Option<Arc<ProjectState>>,
         reservoir_counters: Arc<Mutex<BTreeMap<RuleId, i64>>>,
     ) -> Result<ProcessingStateResult, ProcessingError> {
         // Get the group from the managed envelope context, and if it's not set, try to guess it
@@ -1940,8 +1943,8 @@ impl EnvelopeProcessorService {
     ) -> Result<ProcessEnvelopeResponse, ProcessingError> {
         let ProcessEnvelope {
             envelope: mut managed_envelope,
-            project_info,
-            sampling_project_info,
+            project_state,
+            sampling_project_state,
             reservoir_counters,
         } = message;
 
@@ -1951,7 +1954,7 @@ impl EnvelopeProcessorService {
         //
         // Neither ID can be available in proxy mode on the /store/ endpoint. This is not supported,
         // since we cannot process an envelope without project ID, so drop it.
-        let project_id = match project_info
+        let project_id = match project_state
             .project_id
             .or_else(|| managed_envelope.envelope().meta().project_id())
         {
@@ -1988,8 +1991,8 @@ impl EnvelopeProcessorService {
                 match self.process_envelope(
                     managed_envelope,
                     project_id,
-                    project_info,
-                    sampling_project_info,
+                    project_state,
+                    sampling_project_state,
                     reservoir_counters,
                 ) {
                     Ok(mut state) => {
@@ -2310,7 +2313,7 @@ impl EnvelopeProcessorService {
     fn rate_limit_buckets(
         &self,
         scoping: Scoping,
-        project_state: &ProjectInfo,
+        project_state: &ProjectState,
         mut buckets: Vec<Bucket>,
     ) -> Vec<Bucket> {
         let Some(rate_limiter) = self.inner.rate_limiter.as_ref() else {
@@ -2501,12 +2504,13 @@ impl EnvelopeProcessorService {
 
         let CardinalityLimitsSplit { accepted, rejected } = limits.into_split();
 
-        if !rejected.is_empty() {
-            self.inner
-                .metric_outcomes
-                .track(scoping, &rejected, Outcome::CardinalityLimited);
+        for (bucket, exceeded) in rejected {
+            self.inner.metric_outcomes.track(
+                scoping,
+                &[bucket],
+                Outcome::CardinalityLimited(exceeded.id.clone()),
+            );
         }
-
         accepted
     }
 
@@ -2524,19 +2528,19 @@ impl EnvelopeProcessorService {
         for (scoping, message) in message.scopes {
             let ProjectMetrics {
                 buckets,
-                project_info,
+                project_state,
             } = message;
 
-            let buckets = self.rate_limit_buckets(scoping, &project_info, buckets);
+            let buckets = self.rate_limit_buckets(scoping, &project_state, buckets);
 
-            let limits = project_info.get_cardinality_limits();
+            let limits = project_state.get_cardinality_limits();
             let buckets = self.cardinality_limit_buckets(scoping, limits, buckets);
 
             if buckets.is_empty() {
                 continue;
             }
 
-            let retention = project_info
+            let retention = project_state
                 .config
                 .event_retention
                 .unwrap_or(DEFAULT_EVENT_RETENTION);
@@ -2810,18 +2814,9 @@ impl Service for EnvelopeProcessorService {
     type Interface = EnvelopeProcessor;
 
     fn spawn_handler(self, mut rx: relay_system::Receiver<Self::Interface>) {
-        // Adjust thread count for small cpu counts to not have too many idle cores
-        // and distribute workload better.
-        let thread_count = match self.inner.config.cpu_concurrency() {
-            conc @ 0..=2 => conc.max(1),
-            conc @ 3..=4 => conc - 1,
-            conc => conc - 2,
-        };
-        relay_log::info!("starting {thread_count} envelope processing workers");
+        let semaphore = Arc::new(Semaphore::new(self.inner.pool.current_num_threads()));
 
         tokio::spawn(async move {
-            let semaphore = Arc::new(Semaphore::new(thread_count));
-
             loop {
                 let next_msg = async {
                     let permit_result = semaphore.clone().acquire_owned().await;
@@ -2836,7 +2831,7 @@ impl Service for EnvelopeProcessorService {
 
                     (Some(message), Ok(permit)) = next_msg => {
                         let service = self.clone();
-                        tokio::task::spawn_blocking(move || {
+                        self.inner.pool.spawn(move || {
                             service.handle_message(message);
                             drop(permit);
                         });
@@ -3254,7 +3249,7 @@ mod tests {
         let not_ratelimited_org = 2;
 
         let message = {
-            let project_info = {
+            let project_state = {
                 let quota = Quota {
                     id: Some("testing".into()),
                     categories: vec![DataCategory::MetricBucket].into(),
@@ -3269,10 +3264,9 @@ mod tests {
                 let mut config = ProjectConfig::default();
                 config.quotas.push(quota);
 
-                Arc::new(ProjectInfo {
-                    config,
-                    ..Default::default()
-                })
+                let mut project_state = ProjectState::allowed();
+                project_state.config = config;
+                Arc::new(project_state)
             };
 
             let project_metrics = ProjectMetrics {
@@ -3284,7 +3278,7 @@ mod tests {
                     width: 10,
                     metadata: BucketMetadata::default(),
                 }],
-                project_info,
+                project_state,
             };
 
             let scoping_by_org_id = |org_id: u64| Scoping {
@@ -3391,10 +3385,8 @@ mod tests {
             ..Default::default()
         };
 
-        let project_info = ProjectInfo {
-            config,
-            ..Default::default()
-        };
+        let mut project_state = ProjectState::allowed();
+        project_state.config = config;
 
         let mut envelopes = ProcessingGroup::split_envelope(*envelope);
         assert_eq!(envelopes.len(), 1);
@@ -3404,8 +3396,8 @@ mod tests {
 
         let message = ProcessEnvelope {
             envelope,
-            project_info: Arc::new(project_info),
-            sampling_project_info: None,
+            project_state: Arc::new(project_state),
+            sampling_project_state: None,
             reservoir_counters: ReservoirCounters::default(),
         };
 
@@ -3466,8 +3458,8 @@ mod tests {
 
         let process_message = ProcessEnvelope {
             envelope: managed_envelope,
-            project_info: Arc::new(ProjectInfo::default()),
-            sampling_project_info: None,
+            project_state: Arc::new(ProjectState::allowed()),
+            sampling_project_state: None,
             reservoir_counters: ReservoirCounters::default(),
         };
 
@@ -3825,8 +3817,8 @@ mod tests {
 
     #[test]
     fn test_extrapolate() {
-        let mut project_info = ProjectInfo::default();
-        project_info.config.metric_extraction = ErrorBoundary::Ok(MetricExtractionConfig {
+        let mut project_state = ProjectState::allowed();
+        project_state.config.metric_extraction = ErrorBoundary::Ok(MetricExtractionConfig {
             extrapolate: ExtrapolationConfig {
                 include: vec![LazyGlob::new("*")],
                 exclude: vec![],
@@ -3845,7 +3837,7 @@ mod tests {
         global_config.options.extrapolation_duplication_limit = 3;
 
         let mut extracted_metrics = ProcessingExtractedMetrics::new(
-            Arc::new(project_info),
+            Arc::new(project_state),
             Arc::new(global_config),
             Some(&dsc),
         );