+++ relay-server/src/services/processor/report.rs
@@ -268,7 +268,7 @@ mod tests {
     use crate::extractors::RequestMeta;
     use crate::services::outcome::RuleCategory;
     use crate::services::processor::{ProcessEnvelope, ProcessingGroup};
-    use crate::services::project::ProjectInfo;
+    use crate::services::project::ProjectState;
     use crate::testutils::{self, create_test_processor};
     use crate::utils::ManagedEnvelope;
 
@@ -318,8 +318,8 @@ mod tests {
         let envelope = ManagedEnvelope::new(envelope, outcome_aggregator, test_store, group);
         let message = ProcessEnvelope {
             envelope,
-            project_info: Arc::new(ProjectInfo::default()),
-            sampling_project_info: None,
+            project_state: Arc::new(ProjectState::allowed()),
+            sampling_project_state: None,
             reservoir_counters: ReservoirCounters::default(),
         };
 
@@ -372,8 +372,8 @@ mod tests {
 
         let message = ProcessEnvelope {
             envelope,
-            project_info: Arc::new(ProjectInfo::default()),
-            sampling_project_info: None,
+            project_state: Arc::new(ProjectState::allowed()),
+            sampling_project_state: None,
             reservoir_counters: ReservoirCounters::default(),
         };
 
@@ -434,8 +434,8 @@ mod tests {
         let envelope = ManagedEnvelope::new(envelope, outcome_aggregator, test_store, group);
         let message = ProcessEnvelope {
             envelope,
-            project_info: Arc::new(ProjectInfo::default()),
-            sampling_project_info: None,
+            project_state: Arc::new(ProjectState::allowed()),
+            sampling_project_state: None,
             reservoir_counters: ReservoirCounters::default(),
         };
 
@@ -474,8 +474,8 @@ mod tests {
         let envelope = ManagedEnvelope::new(envelope, outcome_aggregator, test_store, group);
         let message = ProcessEnvelope {
             envelope,
-            project_info: Arc::new(ProjectInfo::default()),
-            sampling_project_info: None,
+            project_state: Arc::new(ProjectState::allowed()),
+            sampling_project_state: None,
             reservoir_counters: ReservoirCounters::default(),
         };
 
@@ -522,8 +522,8 @@ mod tests {
 
         let message = ProcessEnvelope {
             envelope,
-            project_info: Arc::new(ProjectInfo::default()),
-            sampling_project_info: None,
+            project_state: Arc::new(ProjectState::allowed()),
+            sampling_project_state: None,
             reservoir_counters: ReservoirCounters::default(),
         };