+++ relay-server/src/utils/mod.rs
@@ -11,6 +11,7 @@ mod sizes;
 mod sleep_handle;
 mod split_off;
 mod statsd;
+mod thread_pool;
 
 mod memory;
 #[cfg(feature = "processing")]
@@ -36,5 +37,6 @@ pub use self::sizes::*;
 pub use self::sleep_handle::*;
 pub use self::split_off::*;
 pub use self::statsd::*;
+pub use self::thread_pool::*;
 #[cfg(feature = "processing")]
 pub use self::unreal::*;